Server Scripts/requirements.txt
discord.py
Flask
Flask-Limiter
python-dotenv
sqlite3